test installation
This reintroduces pip install to travis, but as a test rather than as
dependency resolution. I'm sure there are more sophisticated ways to
test the install process, but this is practical and easy.

I have intentionally put it at the end of all tests, so we don't
accidentally reintroduce the problem we just fixed: using the installed
code rather than in-tree.